Issue #1000 - Split out common browser functions
